// js/stat-evaluation-player/src/generated/OneTimerEvent.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { RemoteIdTs } from "./RemoteIdTs.ts";

export type OneTimerEvent = { time: number, frame: number, player: RemoteIdTs, player_position?: [number, number, number] | null, passer: RemoteIdTs, passer_position?: [number, number, number] | null, is_team_0: boolean, pass_start_time: number, pass_start_frame: number, pass_duration: number, pass_travel_distance: number, pass_advance_distance: number, ball_speed: number, goal_alignment: number, };