// js/stat-evaluation-player/src/generated/OneTimerPlayerStats.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.

export type OneTimerPlayerStats = { count: number, total_ball_speed: number, fastest_ball_speed: number, total_pass_distance: number, is_last_one_timer: boolean, last_one_timer_time: number | null, last_one_timer_frame: number | null, time_since_last_one_timer: number | null, frames_since_last_one_timer: number | null, };